refactor: add React Query, modals, and notifications
This commit adds React Query, modals, and notifications to the project.

React Query is a library for state management and data fetching. It provides a number of benefits, including:

- Reduced boilerplate code
- Improved performance
- Easier debugging

Modals are a type of dialog box that is used to display information or collect input from the user. They can be used for a variety of purposes, such as:

- Displaying error messages
- Collecting user input
- Displaying additional information

Notifications are a type of message that is displayed to the user for a short period of time. They can be used to:

- Provide feedback to the user
- Display important information
- Alert the user to potential issues