feat: Add multiple private routes to middleware
This change adds support for multiple private routes to the middleware.

Previously, only one private route could be configured, which was "/my-profile". This change allows for multiple private routes to be defined, by adding them to the `privateRoute` array in the middleware.

This allows for more flexibility in configuring which routes require authentication, and makes it easier to protect multiple pages in your application.