fix(layout navbar): add order-history route
Update the onClick function in the Navbar component to push the user to the order-history page when they click on the 'Cek Pesanan' link.

Add a page for displaying the user's order history.
This page will display a table of the user's orders, including the order date, order number, and total price.
The user will be able to click on an order to view more details.